+++ CHANGELOG.md
@@ -13,6 +13,7 @@ changes.
 ### Added
 
 - Add share DRep button to every DRep instead of only our own [Issue 2686](https://github.com/IntersectMBO/govtool/issues/2686)
+- Show metadata anchor in Governance Action Details [Issue 2178](https://github.com/IntersectMBO/govtool/issues/2178)
 
 ### Fixed